# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Humain Numérique - Digital Human AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type MessageInputModel = typeof MessageInputModel[keyof typeof MessageInputModel];


export const MessageInputModel = {
  groq: 'groq',
  gemini: 'gemini',
} as const;

export interface MessageInput {
  content: string;
  /** @nullable */
  conversationId?: string | null;
  model?: MessageInputModel;
  /** @nullable */
  systemPrompt?: string | null;
}

export interface MessageResponse {
  id: string;
  content: string;
  role: string;
  model?: string;
  conversationId?: string;
  createdAt: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  /** @nullable */
  model?: string | null;
  conversationId?: string;
  fromCache?: boolean;
  createdAt: string;
}

export interface Conversation {
  id: string;
  title: string;
  model?: string;
  messageCount?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface ConversationInput {
  title?: string;
  model?: string;
}

export type ChatStatsTopModelsItem = {
  model?: string;
  count?: number;
};

export interface ChatStats {
  totalMessages: number;
  totalConversations: number;
  cachedResponses: number;
  apiCalls: number;
  tokensUsed?: number;
  topModels?: ChatStatsTopModelsItem[];
}

export interface MemoryEntry {
  id: string;
  prompt: string;
  response: string;
  model?: string;
  accessCount?: number;
  createdAt: string;
}

export interface Voice {
  id: string;
  name: string;
  language: string;
  gender: string;
  description?: string;
  /** @nullable */
  preview?: string | null;
}

export interface TtsInput {
  text: string;
  voiceId: string;
  rate?: string;
  pitch?: string;
}

export interface TtsResponse {
  audioUrl: string;
  duration: number;
  text?: string;
}

export type PdfInputFormat = typeof PdfInputFormat[keyof typeof PdfInputFormat];


export const PdfInputFormat = {
  A4: 'A4',
  letter: 'letter',
} as const;

export interface PdfInput {
  title: string;
  content: string;
  format?: PdfInputFormat;
}

export type ZipInputFilesItem = {
  filename: string;
  content: string;
};

export interface ZipInput {
  name: string;
  files: ZipInputFilesItem[];
}

export type FileResultType = typeof FileResultType[keyof typeof FileResultType];


export const FileResultType = {
  pdf: 'pdf',
  zip: 'zip',
} as const;

export interface FileResult {
  id: string;
  name: string;
  type: FileResultType;
  downloadUrl: string;
  size: number;
  createdAt: string;
}

export interface GithubUser {
  login: string;
  /** @nullable */
  name: string | null;
  avatarUrl: string;
  publicRepos: number;
  /** @nullable */
  bio?: string | null;
  followers?: number;
}

export interface GithubRepo {
  id: number;
  name: string;
  fullName: string;
  /** @nullable */
  description?: string | null;
  htmlUrl: string;
  private: boolean;
  /** @nullable */
  language?: string | null;
  stars?: number;
  createdAt: string;
}

export interface GithubRepoInput {
  name: string;
  description?: string;
  private?: boolean;
  autoInit?: boolean;
}

export interface VercelProject {
  id: string;
  name: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  framework?: string | null;
  createdAt: string;
}

export type VercelDeployInputFilesItem = {
  file: string;
  data: string;
};

export interface VercelDeployInput {
  name: string;
  files: VercelDeployInputFilesItem[];
}

export type VercelDeploymentState = typeof VercelDeploymentState[keyof typeof VercelDeploymentState];


export const VercelDeploymentState = {
  BUILDING: 'BUILDING',
  READY: 'READY',
  ERROR: 'ERROR',
  CANCELED: 'CANCELED',
} as const;

export interface VercelDeployment {
  id: string;
  url: string;
  state: VercelDeploymentState;
  name?: string;
  createdAt: string;
  /** @nullable */
  inspectorUrl?: string | null;
}

export type SettingsDefaultModel = typeof SettingsDefaultModel[keyof typeof SettingsDefaultModel];


export const SettingsDefaultModel = {
  groq: 'groq',
  gemini: 'gemini',
} as const;

export type SettingsFontSize = typeof SettingsFontSize[keyof typeof SettingsFontSize];


export const SettingsFontSize = {
  small: 'small',
  medium: 'medium',
  large: 'large',
} as const;

export interface Settings {
  defaultModel: SettingsDefaultModel;
  defaultVoiceId: string;
  streamingEnabled: boolean;
  autoSave: boolean;
  systemPrompt?: string;
  fontSize?: SettingsFontSize;
}

export type SettingsInputDefaultModel = typeof SettingsInputDefaultModel[keyof typeof SettingsInputDefaultModel];


export const SettingsInputDefaultModel = {
  groq: 'groq',
  gemini: 'gemini',
} as const;

export type SettingsInputFontSize = typeof SettingsInputFontSize[keyof typeof SettingsInputFontSize];


export const SettingsInputFontSize = {
  small: 'small',
  medium: 'medium',
  large: 'large',
} as const;

export interface SettingsInput {
  defaultModel?: SettingsInputDefaultModel;
  defaultVoiceId?: string;
  streamingEnabled?: boolean;
  autoSave?: boolean;
  systemPrompt?: string;
  fontSize?: SettingsInputFontSize;
}

export type GetChatHistoryParams = {
conversationId?: string;
limit?: number;
};

export type ClearChatHistoryParams = {
conversationId?: string;
};

export type ListMemoryParams = {
limit?: number;
};

export type SearchMemoryParams = {
q: string;
};