# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Humain Numérique - Digital Human AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/send:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and get AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get conversation history
      parameters:
        - name: conversationId
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear conversation history
      parameters:
        - name: conversationId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /chat/conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /chat/conversations/{id}:
    delete:
      operationId: deleteConversation
      tags: [chat]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        conversationId:
          type: ["string", "null"]
        model:
          type: string
          enum: [groq, gemini]
          default: groq
        systemPrompt:
          type: ["string", "null"]
    MessageResponse:
      type: object
      required: [id, content, role, createdAt]
      properties:
        id:
          type: string
        content:
          type: string
        role:
          type: string
        model:
          type: string
        conversationId:
          type: string
        createdAt:
          type: string
    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        model:
          type: ["string", "null"]
        conversationId:
          type: string
        fromCache:
          type: boolean
        createdAt:
          type: string
    Conversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        model:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ConversationInput:
      type: object
      properties:
        title:
          type: string
        model:
          type: string
...