# lib/api-zod/src/generated/types/conversationInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Humain Numérique - Digital Human AI API
 * OpenAPI spec version: 0.1.0
 */

export interface ConversationInput {
  title?: string;
  model?: string;
}