# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Humain Numérique - Digital Human AI API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageInputModel } from './messageInputModel';

export interface MessageInput {
  content: string;
  /** @nullable */
  conversationId?: string | null;
  model?: MessageInputModel;
  /** @nullable */
  systemPrompt?: string | null;
}